+++ translators/mysql/src/main/groovy/com/virtusa/gto/nyql/db/mysql/MySqlFunctions.groovy
@@ -20,6 +20,28 @@ abstract class MySqlFunctions extends AbstractSQLTranslator implements QFunction
         super(theOptions)
     }
 
+    String greatest(cx) {
+        def c = ___val(cx)
+        def pmx = ___pm(cx)
+
+        if (c instanceof List) {
+            return 'GREATEST' + ___resolveIn(c, pmx)
+        } else {
+            throw new NySyntaxException('GREATEST function requires at least two or more values!')
+        }
+    }
+
+    String least(cx) {
+        def c = ___val(cx)
+        def pmx = ___pm(cx)
+
+        if (c instanceof List) {
+            return 'LEAST' + ___resolveIn(c, pmx)
+        } else {
+            throw new NySyntaxException('LEAST function requires at least two or more values!')
+        }
+    }
+
     @CompileStatic
     @Override
     String truncate(Object cx) {
@@ -99,14 +121,27 @@ abstract class MySqlFunctions extends AbstractSQLTranslator implements QFunction
 
     @CompileStatic
     @Override
-    String cast_to_str(Object col) {
-        String.format('CAST(%s AS CHAR)', ___resolveInP(col))
+    String cast_to_str(Object cx) {
+        def c = ___val(cx)
+        def pmx = ___pm(cx)
+
+        if (c instanceof List && c.size() > 1) {
+            String.format('CAST(%s AS CHAR(%s))', ___resolveIn(c.get(0), pmx), ___resolveIn(c.get(1), pmx))
+        } else {
+            String.format('CAST(%s AS CHAR)', ___resolveInP(cx))
+        }
     }
 
     @CompileStatic
     @Override
     String cast_to_date(Object col) {
         String.format('CAST(%s AS DATE)', ___resolveInP(col))
+    }
+
+    @CompileStatic
+    @Override
+    String cast_to_bigint(Object col) {
+        String.format('CAST(%s AS UNSIGNED INTEGER)', ___resolveInP(col))
     }
 
     @CompileStatic